Refactor : 클래스 명 정리, 불필요한 클래스 삭제 #74
