--- src/main/java/com/sparta/spartatigers/domain/liveboard/service/LiveBoardService.java
@@ -1,96 +1,0 @@
-package com.sparta.spartatigers.domain.liveboard.service;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import com.sparta.spartatigers.domain.liveboard.dto.response.LiveBoardRoomResponseDto;
-import com.sparta.spartatigers.domain.liveboard.model.LiveBoardRoom;
-import com.sparta.spartatigers.domain.liveboard.pubsub.RedisMessageSubscriber;
-import com.sparta.spartatigers.domain.liveboard.repository.LiveBoardRoomRepository;
-import com.sparta.spartatigers.domain.match.model.entity.Match;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class LiveBoardService {
-
-    private final LiveBoardRoomRepository roomRepository;
-    private final RedisMessageListenerContainer redisMessageListener;
-    private final ObjectMapper objectMapper;
-    private final RedisTemplate<String, String> redisTemplate;
-    // private final MatchRepository matchRepository; // TODO: 경기일정 크롤러 확인하기 + 스케줄러
-
-    private Map<String, ChannelTopic> topics = new ConcurrentHashMap<>(); // 채팅방별 topic을 roomId로 찾기
-
-    public void registerRoomTopic(String roomId, SimpMessageSendingOperations messagingTemplate) {
-        if (!topics.containsKey(roomId)) {
-            ChannelTopic topic = new ChannelTopic(roomId);
-            RedisMessageSubscriber subscriber =
-                    new RedisMessageSubscriber(objectMapper, redisTemplate, messagingTemplate);
-            redisMessageListener.addMessageListener(subscriber, topic);
-            topics.put(roomId, topic);
-        }
-    }
-
-    // 입장할 때 호출하면 됨
-    public void enterRoom(String roomId, SimpMessageSendingOperations messagingTemplate) {
-        registerRoomTopic(roomId, messagingTemplate);
-    }
-
-    // 채팅방 전체 조회
-    public List<LiveBoardRoomResponseDto> findAllRoom() {
-        return roomRepository.findAllRoom().stream().map(LiveBoardRoomResponseDto::of).toList();
-    }
-
-    // 채널 토픽 반환 (publish시 사용)
-    public ChannelTopic getTopic(String roomId) {
-        return topics.get(roomId);
-    }
-
-    // 채팅방 생성
-    public void createTodayRoom(List<Match> matches) {
-        for (Match match : matches) {
-            String roomId = "ROOM_" + match.getId();
-            String title = match.getAwayTeam().getName() + "VS" + match.getHomeTeam().getName();
-            LocalDateTime roomOpen = match.getMatchTime().minusMinutes(30);
-            LocalDateTime roomClose = match.getMatchTime().plusHours(5).plusMinutes(30);
-
-            LiveBoardRoom room =
-                    LiveBoardRoom.builder()
-                            .roomId(roomId)
-                            .matchId(match.getId())
-                            .title(title)
-                            .openAt(roomOpen)
-                            .closedAt(roomClose)
-                            .connectCount(0)
-                            .build();
-
-            roomRepository.saveRoom(room);
-        }
-    }
-
-    public void increaseConnectCount(String roomId) {
-        LiveBoardRoom room = roomRepository.findRoomById(roomId);
-        room.increaseCount();
-        roomRepository.saveRoom(room);
-    }
-
-    public void decreaseConnectCount(String roomId) {
-        LiveBoardRoom room = roomRepository.findRoomById(roomId);
-        room.decreaseCount();
-        roomRepository.saveRoom(room);
-    }
-}
